+++ drl_grasping/envs/tasks/reach/reach_octree.py
@@ -48,7 +48,7 @@ class ReachOctree(Reach, abc.ABC):
         )
 
         # Get exact name substitution of the frame for octree
-        octree_reference_frame_id = self.substitute_special_frames(
+        octree_reference_frame_id = self.substitute_special_frame(
             octree_reference_frame_id
         )